[REF] web: convert form compiler tests to hoot
The legacy file contained both compiler and renderer tests.
This commit split those tests into two suites.

task-id: 3705027
